quieten compiler warnings
Misc type coercion issues. Many related to the change of the obs_t SNR
from an int to a float that generate warnings with MSC.

rtksvrostat: change the SNR output argument from int to double,
and rework the callers. The obs_t SNR is now a float rather than an
integer.

str2time: The guard for the width was wrong. Change the start and width
arguments to be size_t to help avoid compiler warnings.

+++ src/rcv/comnav.c
@@ -387,7 +387,7 @@ static int decode_rangeb(raw_t *raw)
             raw->obs.data[index].L  [pos]=-adr;
             raw->obs.data[index].P  [pos]=psr;
             raw->obs.data[index].D  [pos]=(float)dop;
-            raw->obs.data[index].SNR[pos]=0.0<=snr&&snr<255.0?snr:0;
+            raw->obs.data[index].SNR[pos]=0.0<=snr&&snr<255.0?(float)snr:0.0f;
             raw->obs.data[index].LLI[pos]=(unsigned char)lli;
             raw->obs.data[index].code[pos]=code;
 #ifdef RTK_DISABLED
